requirements.txt
blinker==1.6.2
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
Flask==2.3.3
fonttools==4.42.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib2tikz==0.7.6
mysql-connector-python==8.1.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.21.12
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
schedule==1.2.0
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.7
